Create duplicate action file
